+++ Athena-Driver/BSP/Components/DW3000/inc/dw3000deck_ll.h
@@ -18,6 +18,10 @@
 
 #define DUMMY_BYTE 0x00
 
+extern SemaphoreHandle_t spiDeckTxComplete;
+extern SemaphoreHandle_t spiDeckRxComplete;
+extern SemaphoreHandle_t spiDeckMutex;
+
 void BSP_DW3000_Read_ID(uint32_t *dw3000ID);
