+++ Athena-Driver/BSP/Components/DW3000/src/dw3000deck_ll.c
@@ -13,9 +13,9 @@
 #define DW3000Deck_Enable()          LL_GPIO_ResetOutputPin(DW3000Deck_CS_GPIO_Port, DW3000Deck_CS_Pin)
 #define DW3000Deck_Disable()         LL_GPIO_SetOutputPin(DW3000Deck_CS_GPIO_Port, DW3000Deck_CS_Pin)
 
-#define SPI_BUFFER_MAX_SIZE 240
-static uint8_t spiTxBuffer[SPI_BUFFER_MAX_SIZE];
-static uint8_t spiRxBuffer[SPI_BUFFER_MAX_SIZE];
+#define SPI_DECK_BUFFER_MAX_SIZE 240
+static uint8_t spiDeckTxBuffer[SPI_DECK_BUFFER_MAX_SIZE];
+static uint8_t spiDeckRxBuffer[SPI_DECK_BUFFER_MAX_SIZE];
 
 void spiDeckRead(const void* cmd,
 			size_t cmdLength,
@@ -24,10 +24,10 @@ void spiDeckRead(const void* cmd,
 {
 	spiDeckBeginTransaction();
 	DW3000Deck_Enable();
-	memcpy(spiTxBuffer, cmd, cmdLength);
-	memset(spiTxBuffer + cmdLength, DUMMY_BYTE, dataLength);
-	spiDeckExchange(cmdLength + dataLength, spiTxBuffer, spiRxBuffer);
-	memcpy(data, spiRxBuffer + cmdLength, dataLength);
+	memcpy(spiDeckTxBuffer, cmd, cmdLength);
+	memset(spiDeckTxBuffer + cmdLength, DUMMY_BYTE, dataLength);
+	spiDeckExchange(cmdLength + dataLength, spiDeckTxBuffer, spiDeckRxBuffer);
+	memcpy(data, spiDeckRxBuffer + cmdLength, dataLength);
 	DW3000Deck_Disable();
 	spiDeckEndTransaction();
 }
@@ -47,8 +47,8 @@ void BSP_DW3000_Read_ID(uint32_t *dw3000ID)
 	 * - spiRead(headerBuffer, headerLength, readBuffer, readlength);
 	 */
 
-	uint8_t cmd[4] = {0x40, DUMMY_BYTE, DUMMY_BYTE, DUMMY_BYTE};
-	led_flash_delay_in_ms = 4000;
-	spiDeckRead(cmd, 4, dw3000ID, 4);
+	uint8_t cmd[4] = {0x0, DUMMY_BYTE, DUMMY_BYTE, DUMMY_BYTE};
+	led_flash_delay_in_ms = 3000;
+	spiDeckRead(0, 1, dw3000ID, 4);
 	return ;
 }
