+++ Core/Inc/spi.h
@@ -35,9 +35,6 @@ extern "C" {
 /* USER CODE END Includes */
 
 /* USER CODE BEGIN Private defines */
-extern SemaphoreHandle_t spiDeckTxComplete;
-extern SemaphoreHandle_t spiDeckRxComplete;
-extern SemaphoreHandle_t spiDeckMutex;
 
 /* USER CODE END Private defines */
