+++ Core/Src/stm32l4xx_it.c
@@ -27,7 +27,7 @@
 #include "semphr.h"
 #include "task.h"
 #include "spi_drv.h"
-#include "spi.h"
+#include "dw3000deck_ll.h"
 #include "uart_receive.h"
 /* USER CODE END Includes */
 
@@ -221,8 +221,8 @@ void DMA1_Channel4_IRQHandler(void)
 {
   /* USER CODE BEGIN DMA1_Channel4_IRQn 0 */
 	BaseType_t xHigherPriorityTaskWoken = pdFALSE;
-	if (LL_DMA_IsActiveFlag_TC1(DMA1)){
-		LL_DMA_ClearFlag_TC1(DMA1);
+	if (LL_DMA_IsActiveFlag_TC4(DMA1)){
+		LL_DMA_ClearFlag_TC4(DMA1);
 		LL_SPI_DisableDMAReq_RX(SPI2);
 		LL_DMA_DisableChannel(DMA1, LL_DMA_CHANNEL_4);
 		xSemaphoreGiveFromISR(spiDeckRxComplete, &xHigherPriorityTaskWoken);
@@ -243,8 +243,8 @@ void DMA1_Channel5_IRQHandler(void)
 {
   /* USER CODE BEGIN DMA1_Channel5_IRQn 0 */
 	BaseType_t xHigherPriorityTaskWoken = pdFALSE;
-	if (LL_DMA_IsActiveFlag_TC1(DMA1)){
-		LL_DMA_ClearFlag_TC1(DMA1);
+	if (LL_DMA_IsActiveFlag_TC5(DMA1)){
+		LL_DMA_ClearFlag_TC5(DMA1);
 		LL_SPI_DisableDMAReq_TX(SPI2);
 		LL_DMA_DisableChannel(DMA1, LL_DMA_CHANNEL_5);
 		xSemaphoreGiveFromISR(spiDeckTxComplete, &xHigherPriorityTaskWoken);
