+++ athena-efficiency.ioc
@@ -224,7 +224,7 @@ NVIC.USART1_IRQn=true\:5\:0\:false\:false\:true\:true\:true\:true\:true
 NVIC.UsageFault_IRQn=true\:0\:0\:false\:false\:true\:false\:false\:false\:false
 PA1.GPIOParameters=GPIO_Speed,PinState,GPIO_PuPd
 PA1.GPIO_PuPd=GPIO_PULLUP
-PA1.GPIO_Speed=GPIO_SPEED_FREQ_HIGH
+PA1.GPIO_Speed=GPIO_SPEED_FREQ_MEDIUM
 PA1.Locked=true
 PA1.PinState=GPIO_PIN_SET
 PA1.Signal=GPIO_Output
@@ -327,7 +327,7 @@ ProjectManager.ToolChainLocation=
 ProjectManager.UAScriptAfterPath=
 ProjectManager.UAScriptBeforePath=
 ProjectManager.UnderRoot=true
-ProjectManager.functionlistsort=1-SystemClock_Config-RCC-false-LL-false,2-MX_GPIO_Init-GPIO-false-LL-true,3-MX_DMA_Init-DMA-false-LL-true,4-MX_I2C1_Init-I2C1-false-LL-true,5-MX_I2C3_Init-I2C3-false-LL-true,6-MX_SPI3_Init-SPI3-false-LL-true,7-MX_SPI2_Init-SPI2-false-LL-true,8-MX_SPI1_Init-SPI1-false-LL-true,9-MX_USART1_UART_Init-USART1-false-LL-true,10-MX_USART3_UART_Init-USART3-false-LL-true
+ProjectManager.functionlistsort=1-SystemClock_Config-RCC-false-LL-false,2-MX_GPIO_Init-GPIO-false-LL-true,3-MX_DMA_Init-DMA-false-LL-true,4-MX_I2C1_Init-I2C1-false-LL-true,5-MX_I2C3_Init-I2C3-false-LL-true,6-MX_SPI3_Init-SPI3-false-LL-true,7-MX_SPI2_Init-SPI2-false-LL-true,8-MX_SPI1_Init-SPI1-false-LL-true,9-MX_USART1_UART_Init-USART1-false-LL-true
 RCC.ADCFreq_Value=32000000
 RCC.AHBFreq_Value=80000000
 RCC.APB1Freq_Value=80000000
